// ui/src/generated/ObjectObservation.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ObjectMetadata } from "./ObjectMetadata";
import type { ObjectPathElement } from "./ObjectPathElement";

export type ObjectObservation = { object_ref: string, name: string, path: Array<ObjectPathElement>, metadata: ObjectMetadata, observed_at_ms: number, expires_at_ms: number, };